plotTrackOverlay_Dcoef

Plot track/trajectory overlays from a list of files in a folder with color coded by the Diffusion Coefficient (Dcoef) of each track/trajectory.

plotTrackOverlay_Dcoef(trackll=trackll,scale=128,dt=6,
filter=c(min=7,max=Inf),
resolution=0.107,rsquare=0.8,t.interval=0.01,Dcoef.range=c(-6,2),
color=c("blue", "white", "red"),folder=NULL,file.No=0,line.width=0.1)

Arguments

trackll
trajectory list generated by createTrackll() and processing. if NULL, user will be prompted to enter the trackll name.

scale
The pixel scale of image data.

dt
Time inbervals.

filter
Filter the tracks by step/frame number (length). Only tracks pass through filter will be selected.

resolution
ratio of pixel to uM.

rsquare
R square filter on Dcoef results. Default to be 0.8. Set value to 0 if rsquare filter is not desired.

t.interval
Time interval for image aquisition. Default 0.01 sec (10ms).

Dcoef.range
Select tracks whose Dcoef is within the range. Other tracks will not be plotted.

color
colors to interpolate to build the color gradient/ramp.

folder
Path to provide nuclear glow backround. Images should end with "_Nuclei.tif". Default is NULL and not adding background.

file.No
Select file(s) in the folder to plot. Default 0 for plotting all files in the folder.

line.width
Line width for plotted tracks, default is 0.1.

Value

PDF: One PDF file with one plot on each page.

Details

Plot track/trajectory overlays from a list of files in a folder with color coded by the Diffusion Coefficient (Dcoef) of each track/trajectory. The range of Dcoef can be defined. Nuclei backgrounds can be added to the plots. Tracks from each file will be plot in one plot, i.e., one plot per file. Colors are flexible.

Upon running of the function, users will be prompted to input the name of the track list (trackll). Input un-merged trackll and the plotting will start. The Dcoef is calculated by "static" method, which stabilizes the number of time lags used for fitting using time lag 2~ 5 despite the total time lags measured.

Examples

# Generate trackll, and process, 
# e.g. mask region of interest, tracks from multiple files should not be 
# merged.
folder=system.file("extdata","HSF",package="sojourner")
trackll=createTrackll(folder=folder, input=3)

#> 
#> Reading ParticleTracker file:  HSF-Halo_1.csv ...
#> 
#>  alo_1 read and processed.
#> 
#> Reading ParticleTracker file:  HSF-Halo_2.csv ...
#> 
#>  alo_2 read and processed.
#> 
#> Process complete.

trackll=maskTracks(folder,trackll)

#> Reading mask file     HSF-Halo_1_MASK.tif 
#> Reading mask file     HSF-Halo_2_MASK.tif 
#> 
#> All files masked.

# Plot track overlays,
#plotTrackOverlay_Dcoef(trackll=trackll,scale=128,dt=6,
#filter=c(min=7,max=Inf),
#resolution=0.107,rsquare=0.8,t.interval=0.01,Dcoef.range=c(-6,2),
#color=c("blue", "white", "red"),folder=NULL,file.No=0,line.width=0.1)
plotTrackOverlay_Dcoef(trackll=trackll,color=c("red", "yellow"),
folder=folder,file.No=0)

#> applying filter, min 7   max Inf 
#> applying filter, min 7   max Inf 
#> Plotting...A PDF file will be output in the working directory.

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> pdf 
#>   2 

plotTrackOverlay_Dcoef(trackll=trackll,color=c("red", "yellow"),
folder=folder,file.No=c(1,2))

#> applying filter, min 7   max Inf 
#> applying filter, min 7   max Inf 
#> Plotting...A PDF file will be output in the working directory.

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> pdf 
#>   2