readParticleTracker

Read output file (tracks/trajectories in csv format) from ParticleTracker (a program of ImageJ plugin MosaicSuit).

readParticleTracker(folder, ab.track = FALSE, cores = 1, frameRecord = TRUE)

Arguments

ab.track
Use absolute coordinates for tracks.

frameRecord
Add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (especially helpful when linking frames).

folder
Full path to ImageJ .csv files output folder.

cores
Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralleled.

Value

trackll

Details

The usage of readParticleTracker() is equivalent to readDiatrack().

Note: the folder name should not contain ".", as it is a key character for subsequent indexing of file names.

trackID=fileID.frameID.duration.indexPerFile.indexPerTrackll

This "indexPerFile" is the index within a track file.

This "indexPerTrackll" is the index within a trackll, which is unique.

The macro used for generating the csv file is also included in ImageJ folder of the package: folder=system.file("extdata","ImageJ",package="sojourner")

Examples

# reading in tracks
folder=system.file("extdata","ImageJ",package="sojourner")
trackll=readParticleTracker(folder)

#> 
#> Reading ParticleTracker file:  epl11AA_HTZ1Halo_No_RAP_fr10ms_120mW_1.tif.csv ...
#> 
#>  0mW_1 read and processed.
#> 
#> Reading ParticleTracker file:  epl11AA_HTZ1Halo_No_RAP_fr10ms_120mW_2.tif.csv ...
#> 
#>  0mW_2 read and processed.
#> 
#> Process complete.

str(trackll,max.level=2)

#> List of 2
#>  $ epl11AA_HTZ1Halo_No_RAP_fr10ms_120mW_1.tif.csv:List of 800
#>   ..$ 0mW_1.0.10.1.1        :'data.frame':	10 obs. of  4 variables:
#>   ..$ 0mW_1.10.3.2.2        :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.14.3.3.3        :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.20.25.4.4       :'data.frame':	25 obs. of  4 variables:
#>   ..$ 0mW_1.49.2.5.5        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.51.3.6.6        :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.54.16.7.7       :'data.frame':	16 obs. of  4 variables:
#>   ..$ 0mW_1.71.2.8.8        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.73.7.9.9        :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_1.83.2.10.10      :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.87.4.11.11      :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_1.91.15.12.12     :'data.frame':	15 obs. of  4 variables:
#>   ..$ 0mW_1.112.3.13.13     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.119.4.14.14     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_1.128.2.15.15     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.130.2.16.16     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.132.6.17.17     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_1.138.2.18.18     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.141.27.19.19    :'data.frame':	27 obs. of  4 variables:
#>   ..$ 0mW_1.170.2.20.20     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.172.4.21.21     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_1.177.2.22.22     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.180.2.23.23     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.185.3.24.24     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.190.2.25.25     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.193.2.26.26     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.197.5.27.27     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_1.209.2.28.28     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.212.4.29.29     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_1.217.2.30.30     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.225.2.31.31     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.228.3.32.32     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.232.5.33.33     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_1.247.2.34.34     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.252.2.35.35     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.255.2.36.36     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.270.4.37.37     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_1.278.2.38.38     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.283.2.39.39     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.290.3.40.40     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.299.3.41.41     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.314.3.42.42     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.318.2.43.43     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.322.3.44.44     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.327.2.45.45     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.332.2.46.46     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.335.2.47.47     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.339.19.48.48    :'data.frame':	19 obs. of  4 variables:
#>   ..$ 0mW_1.358.2.49.49     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.361.29.50.50    :'data.frame':	29 obs. of  4 variables:
#>   ..$ 0mW_1.394.2.51.51     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.398.2.52.52     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.462.2.53.53     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.468.3.54.54     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.493.2.55.55     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.496.3.56.56     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.503.2.57.57     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.506.2.58.58     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.509.3.59.59     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.512.2.60.60     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.514.7.61.61     :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_1.535.3.62.62     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.540.2.63.63     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.543.2.64.64     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.563.2.65.65     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.568.2.66.66     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.572.2.67.67     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.578.2.68.68     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.583.2.69.69     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.585.2.70.70     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.590.3.71.71     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.599.2.72.72     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.601.3.73.73     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.617.2.74.74     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.636.3.75.75     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.639.13.76.76    :'data.frame':	13 obs. of  4 variables:
#>   ..$ 0mW_1.659.3.77.77     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.663.14.78.78    :'data.frame':	14 obs. of  4 variables:
#>   ..$ 0mW_1.681.5.79.79     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_1.686.9.80.80     :'data.frame':	9 obs. of  4 variables:
#>   ..$ 0mW_1.695.15.81.81    :'data.frame':	15 obs. of  4 variables:
#>   ..$ 0mW_1.700.3.82.82     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.716.2.83.83     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.719.2.84.84     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.720.2.85.85     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.724.2.86.86     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.726.7.87.87     :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_1.734.2.88.88     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.736.8.89.89     :'data.frame':	8 obs. of  4 variables:
#>   ..$ 0mW_1.756.12.90.90    :'data.frame':	12 obs. of  4 variables:
#>   ..$ 0mW_1.770.3.91.91     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.773.6.92.92     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_1.784.3.93.93     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.792.5.94.94     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_1.799.2.95.95     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_1.800.3.96.96     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.804.5.97.97     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_1.813.3.98.98     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_1.816.2.99.99     :'data.frame':	2 obs. of  4 variables:
#>   .. [list output truncated]
#>  $ epl11AA_HTZ1Halo_No_RAP_fr10ms_120mW_2.tif.csv:List of 544
#>   ..$ 0mW_2.0.10.1.1        :'data.frame':	10 obs. of  4 variables:
#>   ..$ 0mW_2.14.3.2.2        :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.21.2.3.3        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.25.2.4.4        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.28.2.5.5        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.33.5.6.6        :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_2.55.2.7.7        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.57.2.8.8        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.60.2.9.9        :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.63.2.10.10      :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.68.3.11.11      :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.71.2.12.12      :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.89.4.13.13      :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.94.2.14.14      :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.97.2.15.15      :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.103.9.16.16     :'data.frame':	9 obs. of  4 variables:
#>   ..$ 0mW_2.113.7.17.17     :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_2.125.7.18.18     :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_2.130.6.19.19     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_2.138.2.20.20     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.142.5.21.21     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_2.148.2.22.22     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.150.7.23.23     :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_2.158.2.24.24     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.160.6.25.25     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_2.167.6.26.26     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_2.172.20.27.27    :'data.frame':	20 obs. of  4 variables:
#>   ..$ 0mW_2.194.2.28.28     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.196.35.29.29    :'data.frame':	35 obs. of  4 variables:
#>   ..$ 0mW_2.233.10.30.30    :'data.frame':	10 obs. of  4 variables:
#>   ..$ 0mW_2.244.3.31.31     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.247.4.32.32     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.251.2.33.33     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.255.41.34.34    :'data.frame':	41 obs. of  4 variables:
#>   ..$ 0mW_2.299.4.35.35     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.304.73.36.36    :'data.frame':	73 obs. of  4 variables:
#>   ..$ 0mW_2.384.13.37.37    :'data.frame':	13 obs. of  4 variables:
#>   ..$ 0mW_2.400.2.38.38     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.403.2.39.39     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.412.4.40.40     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.418.7.41.41     :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_2.426.2.42.42     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.430.9.43.43     :'data.frame':	9 obs. of  4 variables:
#>   ..$ 0mW_2.443.15.44.44    :'data.frame':	15 obs. of  4 variables:
#>   ..$ 0mW_2.460.4.45.45     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.467.5.46.46     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_2.475.6.47.47     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_2.480.6.48.48     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_2.489.3.49.49     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.491.3.50.50     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.498.10.51.51    :'data.frame':	10 obs. of  4 variables:
#>   ..$ 0mW_2.509.2.52.52     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.513.12.53.53    :'data.frame':	12 obs. of  4 variables:
#>   ..$ 0mW_2.525.4.54.54     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.529.11.55.55    :'data.frame':	11 obs. of  4 variables:
#>   ..$ 0mW_2.549.21.56.56    :'data.frame':	21 obs. of  4 variables:
#>   ..$ 0mW_2.576.2.57.57     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.580.5.58.58     :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_2.587.7.59.59     :'data.frame':	7 obs. of  4 variables:
#>   ..$ 0mW_2.600.87.60.60    :'data.frame':	87 obs. of  4 variables:
#>   ..$ 0mW_2.697.64.61.61    :'data.frame':	64 obs. of  4 variables:
#>   ..$ 0mW_2.774.2.62.62     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.786.3.63.63     :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.793.2.64.64     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.795.2.65.65     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.803.28.66.66    :'data.frame':	28 obs. of  4 variables:
#>   ..$ 0mW_2.831.4.67.67     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.839.11.68.68    :'data.frame':	11 obs. of  4 variables:
#>   ..$ 0mW_2.852.2.69.69     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.872.2.70.70     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.877.2.71.71     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.915.2.72.72     :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.945.4.73.73     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.951.15.74.74    :'data.frame':	15 obs. of  4 variables:
#>   ..$ 0mW_2.971.6.75.75     :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_2.977.14.76.76    :'data.frame':	14 obs. of  4 variables:
#>   ..$ 0mW_2.995.4.77.77     :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.1000.5.78.78    :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_2.1008.2.79.79    :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.1010.2.80.80    :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.1011.4.81.81    :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.1016.3.82.82    :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.1020.3.83.83    :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.1023.3.84.84    :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.1026.3.85.85    :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.1029.4.86.86    :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.1032.2.87.87    :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.1036.3.88.88    :'data.frame':	3 obs. of  4 variables:
#>   ..$ 0mW_2.1039.27.89.89   :'data.frame':	27 obs. of  4 variables:
#>   ..$ 0mW_2.1055.4.90.90    :'data.frame':	4 obs. of  4 variables:
#>   ..$ 0mW_2.1074.38.91.91   :'data.frame':	38 obs. of  4 variables:
#>   ..$ 0mW_2.1125.2.92.92    :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.1127.38.93.93   :'data.frame':	38 obs. of  4 variables:
#>   ..$ 0mW_2.1172.5.94.94    :'data.frame':	5 obs. of  4 variables:
#>   ..$ 0mW_2.1181.2.95.95    :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.1191.2.96.96    :'data.frame':	2 obs. of  4 variables:
#>   ..$ 0mW_2.1197.6.97.97    :'data.frame':	6 obs. of  4 variables:
#>   ..$ 0mW_2.1202.57.98.98   :'data.frame':	57 obs. of  4 variables:
#>   ..$ 0mW_2.1267.2.99.99    :'data.frame':	2 obs. of  4 variables:
#>   .. [list output truncated]