readUtrack

take in a single channel Utrack file as input, along with several other user-configurable parameters and output options, to return a track list of all the trajectories found in the file

readUtrack(folder, ab.track = FALSE, cores = 1, frameRecord = TRUE)

Arguments

ab.track
Use absolute coordinates for tracks.

frameRecord
Add a fourth column to the track list after the xyz-coordinates for the frame that coordinate point was found (almost mandatory for Utrack).

folder
Full path to Utrack files output folder.

cores
Number of cores used for parallel computation. This can be the cores on a workstation, or on a cluster. Tip: each core will be assigned to read in a file when paralleled.

Value

trackll

Details

The naming scheme for each track is as follows:

[Last five characters of the file name].[Start frame #].[Length].[Track #]

(Note: The last five characters of the file name, excluding the extension, cannot contain ".")

Examples

#Basic function call of readUtrack
# trackll <- readUtrack(folder=folder)